requirements/test.txt
-r base.txt
Flask-Testing==0.7.1
pytest==4.0.2
pytest-cov==2.6.0
flake8==3.6.0
responses==0.10.5